+++ CHANGELOG.md
@@ -34,6 +34,7 @@ CrystFEL's `partialator` has the ability to split a processed `.stream` file int
 - Since we now had runs with more than 1 million frames, the max frame limit was increased to 2 million for now, while we are working on a much better solution ([\#504](https://gitlab.desy.de/amarcord/amarcord/-/issues/504))
 - If there are processing errors somewhere inside a data set, this is now indicated on the top of the indexing parameters ([\#506](https://gitlab.desy.de/amarcord/amarcord/-/issues/506)).
 - With the SLURM REST workload manager backend, we now query each running job separately, instead of doing a bulk request. This is much faster, since we don't have to choose a time span to query the jobs, and cannot accidentally "overquery" thousands of jobs, resulting in huge delays (yes, this has happened). ([\#504](https://gitlab.desy.de/amarcord/amarcord/-/issues/504), [\#505](https://gitlab.desy.de/amarcord/amarcord/-/issues/505))
+- In the ID29 case, you can now let the ID29 push daemon create the attributi defined in the config file instead of manually adding them ([\#513](https://gitlab.desy.de/amarcord/amarcord/-/issues/513))
 
 ## Fixes